handwriting-app/requirements.txt
tensorflow
numpy
opencv-python
pillow
emnist
matplotlib
